% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_match_col.R
\name{add_match_col}
\alias{add_match_col}
\title{Create column with matched value}
\usage{
add_match_col(string, pattern)
}
\arguments{
\item{string}{string to search}

\item{pattern}{regex pattern to match}
}
\value{
matched string
}
\description{
Create column with matched value
}
\examples{
require(dplyr)
require(tidyr)
dplyr::mutate(head(dplyr::starwars),
      match = add_match_col(name, "Skywalker")) |>
  dplyr::select(name, last_col())
}
