% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_regex_col}
\alias{add_regex_col}
\title{Match string and add column with matched value}
\usage{
add_regex_col(string, pattern)
}
\arguments{
\item{string}{string to search}

\item{pattern}{regex pattern to match}
}
\value{
matched string
}
\description{
Match string and add column with matched value
}
