% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_substance.R
\name{classify_substance}
\alias{classify_substance}
\title{Classify a specific WADA substance}
\usage{
classify_substance(
  df,
  substance,
  value,
  substance_col = "substance_reason",
  wb = FALSE
)
}
\arguments{
\item{df}{data from USADA website}

\item{substance}{character string of banned WADA substances}

\item{value}{character string of substance group value}

\item{substance_col}{column with substances/reasons for sanctions (assumes
it's \code{substance_reason}).}

\item{wb}{logical, include word boundary?}
}
\value{
substance dataset with newly classified substances
}
\description{
Classify a specific WADA substance
}
