% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_sports.R
\name{clean_sports}
\alias{clean_sports}
\title{Clean/tidy USADA sports}
\usage{
clean_sports(df, sport_col, tidy = TRUE)
}
\arguments{
\item{df}{data.frame/tibble of processed USADA dataset with messy sports}

\item{sport_col}{character sport column (usually \code{sport})}

\item{tidy}{logical. Tidy the sports after cleaning/wrangling?}
}
\value{
tibble with cleaned sports
}
\description{
Clean/tidy USADA sports
}
