% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_top_file.R
\name{clip_top_file}
\alias{clip_top_file}
\title{Copy the path to top data file in \verb{extdata/} folder on the clipboard}
\usage{
clip_top_file(ext, pth, ctime)
}
\arguments{
\item{ext}{file extension. Default is \code{'csv'}, accepts: \code{'rds'}, \code{'txt'}, \code{'tsv'},
or \code{'dat'}}

\item{pth}{path to \code{extdata} subfolder}

\item{ctime}{modification time of the most recent file, as a POSIXct datetime}
}
\value{
Called for its side effects: prints file name and modification time to
the console, and copies a ready-to-use import statement to the clipboard.
Returns the clipboard content string invisibly.
}
\description{
Copy the path to top data file in \verb{extdata/} folder on the clipboard
}
