% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dir_date.R
\name{create_dir_date}
\alias{create_dir_date}
\title{Create a folder (with date-stamped sub-folder)}
\usage{
create_dir_date(dir_path = "", dir_name)
}
\arguments{
\item{dir_path}{parent folder}

\item{dir_name}{directory name}
}
\value{
path to folder
}
\description{
Create a folder (with date-stamped sub-folder)
}
\examples{
create_dir_date(dir_name = "tmp")
create_dir_date(dir_path = "tmp", dir_name = "doc")
create_dir_date(dir_path = "tmp/x", dir_name = "doc")
create_dir_date(dir_path = "tmp/x/", dir_name = "doc")
}
