% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_regex_wb.R
\name{create_regex_wb}
\alias{create_regex_wb}
\title{Create a regular expression with word boundaries}
\usage{
create_regex_wb(string)
}
\arguments{
\item{string}{character vector of items}
}
\value{
wb_regex regular expression
}
\description{
Create a regular expression with word boundaries
}
\examples{
require(stringr)
wb_regex <- create_regex_wb(c("pink", "salmon."))
str_view_all(stringr::sentences, wb_regex, match = TRUE)
}
