% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_word_boundary.R
\name{create_word_boundary}
\alias{create_word_boundary}
\title{Create a regular expression with word boundaries}
\usage{
create_word_boundary(string)
}
\arguments{
\item{string}{character vector of items}
}
\value{
wb_regex regular expression
}
\description{
Create a regular expression with word boundaries
}
\examples{
require(stringr)
wb_regex <- create_word_boundary(c("pink", "salmon."))
str_view(stringr::sentences, wb_regex, match = TRUE)
}
