% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtstamp.R
\name{dtstamp}
\alias{dtstamp}
\title{Insert date/time stamp}
\usage{
dtstamp(include_time = FALSE, side = "none")
}
\arguments{
\item{include_time}{logical, include time?}

\item{side}{include an underscore (_) on the 'left' or 'right' side
(default is 'none')}
}
\value{
polished date (or date and time) stamp
}
\description{
Insert date/time stamp
}
\examples{
dtstamp()
dtstamp(TRUE)
dtstamp(FALSE, "r")
dtstamp(TRUE, "l")
}
