% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_data.R
\name{export_data}
\alias{export_data}
\title{Export data object to path}
\usage{
export_data(x, path = "", type = "csv")
}
\arguments{
\item{x}{dataset to export}

\item{path}{string, path to folder}

\item{type}{string, type of exported file}
}
\value{
exported data message
}
\description{
\code{export_data()} will export a given \code{data.frame} or \code{tibble} to a specified
path with a data stamp prefix.
}
