% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_extdata.R
\name{export_extdata}
\alias{export_extdata}
\title{Export data to \verb{inst/extdata/} or \code{inst/extdata/raw}}
\usage{
export_extdata(x, path = "", type = "csv")
}
\arguments{
\item{x}{data to export}

\item{path}{path to 'inst/' folder (and only 'inst/' folder!)}

\item{type}{logical, is this a raw dataset?}
}
\value{
exported data
}
\description{
Export data to \verb{inst/extdata/} or \code{inst/extdata/raw}
}
