% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sanctions_data.R
\name{get_prohibited_association}
\alias{get_prohibited_association}
\title{Get USADA prohibited-association data}
\usage{
get_prohibited_association(refresh = FALSE, max_age_days = 7, verbose = FALSE)
}
\arguments{
\item{refresh}{force a refresh of the cache before returning data.
Default \code{FALSE}}

\item{max_age_days}{maximum cache age, in days, before an automatic
refresh is triggered. Default \code{7}}

\item{verbose}{print progress messages. Default \code{FALSE}}
}
\value{
a \code{data.frame} of currently-ineligible support personnel
}
\description{
Returns the list of athlete support personnel (coaches, trainers, etc.)
currently ineligible under the Prohibited Association rule (Name,
Suspension Ends), using a local cache that's automatically refreshed when
stale.
}
\examples{
\donttest{
get_prohibited_association()
}
}
