% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recent.R
\name{get_recent}
\alias{get_recent}
\title{Return the most recent modification date}
\usage{
get_recent(pth = ".", full = FALSE)
}
\arguments{
\item{pth}{path to file or folder}

\item{full}{return datetime (instead of date)}
}
\value{
A named character vector: \code{"dstmp"} containing the most recent
modification date, or \code{"dtstmp"} containing the full datetime when
\code{full = TRUE}.
}
\description{
Return the most recent modification date
}
\examples{
\dontrun{
# invalid path raises an error
get_recent("wrong")
}
get_recent(tempdir())
get_recent(tempdir(), full = TRUE)
}
