% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recent_data_file.R
\name{get_recent_data_file}
\alias{get_recent_data_file}
\title{Return current data in folder}
\usage{
get_recent_data_file(folder, type)
}
\arguments{
\item{folder}{directory with data files}

\item{type}{file extension}
}
\value{
path to recent data file
}
\description{
Return current data in folder
}
\examples{
# not run
get_recent_data_file(folder = "inst/extdata/", type = "txt")
get_recent_data_file(folder = "inst/extdata/", type = "csv")
get_recent_data_file(folder = "inst/extdata/", type = "tsv")
}
