% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recent_version.R
\name{get_recent_version}
\alias{get_recent_version}
\title{Return the most recent modification time for folder and files}
\usage{
get_recent_version(pth = ".", full = FALSE)
}
\arguments{
\item{pth}{path to file or folder}

\item{full}{return datetime (instead of date)}
}
\value{
vrsn character vector of date or datetime
}
\description{
Return the most recent modification time for folder and files
}
\examples{
get_recent_version("wrong")
get_recent_version("inst/extdata/")
get_recent_version("inst/extdata/raw", full = TRUE)
}
