% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nin.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Not in (not-in operator for R.)}
\usage{
x \%nin\% y
}
\arguments{
\item{x}{vector or \code{NULL}: the values to be matched.}

\item{y}{vector or \code{NULL}: the values to be matched against.}
}
\value{
The negation of \code{\link[base:match]{\%in\%}}.
}
\description{
Not in (not-in operator for R.)
}
\examples{
1 \%nin\% 2:10
c("a", "b") \%nin\% c("a", "c", "d")
}
