% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polite-scrape.R
\name{\%otherwise\%}
\alias{\%otherwise\%}
\title{null-coalescing operator. See purrr for details.}
\usage{
lhs \%otherwise\% rhs
}
\arguments{
\item{lhs}{left hand side}

\item{rhs}{right hand side}
}
\description{
null-coalescing operator. See purrr for details.
}
