% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polite-scrape.R
\name{polite_download_file}
\alias{polite_download_file}
\title{Polite download}
\usage{
polite_download_file(
  url,
  destfile = guess_basename(url),
  ...,
  quiet = !verbose,
  mode = "wb",
  path = "downloads/",
  user_agent = paste0("polite ", getOption("HTTPUserAgent")),
  delay = 5,
  force = FALSE,
  overwrite = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{url}{web address for the file to be downloaded}

\item{destfile}{name of destination file}

\item{...}{additional arguments passed to \code{download.file}}

\item{quiet}{default value is inverse of  \code{verbose}}

\item{mode}{download mode. Default value is "wb"}

\item{path}{path to save. Default path \verb{downloads/}}

\item{user_agent}{default value \code{paste0("polite ", getOption("HTTPUserAgent"))}}

\item{delay}{default value equal 5}

\item{force}{force re-download of robots.txt}

\item{overwrite}{overwrite downloaded file. Default value FALSE}

\item{verbose}{default value is FALSE}
}
\description{
Polite download
}
\keyword{internal}
