% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polite-scrape.R
\name{polite_read_html}
\alias{polite_read_html}
\title{function that actually fetches response from the web}
\usage{
polite_read_html(
  url,
  ...,
  delay = 5,
  user_agent = paste0("polite ", getOption("HTTPUserAgent"), "bot"),
  force = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{url}{web address for scraping}

\item{...}{arguments passed to \code{httr::GET()}}

\item{delay}{scraping delay. Default 5 sec}

\item{user_agent}{user agent string. Default value \code{paste0("polite ", getOption("HTTPUserAgent"), "bot")}}

\item{force}{force re-download of robots.txt}

\item{verbose}{default FALSE}
}
\description{
function that actually fetches response from the web
}
