% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_text.R
\name{process_text}
\alias{process_text}
\title{Process raw data}
\usage{
process_text(raw_data, fct = FALSE)
}
\arguments{
\item{raw_data}{raw dataset}

\item{fct}{include factors?}
}
\value{
data with standardized names, lowercase text, etc.
}
\description{
Process raw data
}
\examples{
usada_raw <- data.frame(
  Athlete = c("Doe, Jane", "Roe, Richard"),
  Sport = c("Cycling", "Weightlifting"),
  Substance.Reason = c("Cannabinoids", "Ostarine"),
  Sanction.Terms = c("2-Year Suspension", "1-Year Suspension"),
  Sanction.Announced = c("01/01/2024", "02/15/2024")
)
str(usada_raw)
# compare to
str(process_text(usada_raw))
}
