% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_cols.R
\name{split_cols}
\alias{split_cols}
\title{Separate column into multiple columns}
\usage{
split_cols(data, col, pattern = "[^[:alnum:]]+", col_prefix)
}
\arguments{
\item{data}{\code{data.frame} or \code{tibble}}

\item{col}{column to separate}

\item{pattern}{regular expression pattern passed to \code{strsplit()}}

\item{col_prefix}{prefix for new columns}
}
\value{
data.frame with split columns
}
\description{
Separate column into multiple columns
}
\examples{
d <- data.frame(value = c(29L, 91L, 39L, 28L, 12L),
                name = c("John", "John, Jacob",
                         "John, Jacob, Jingleheimer",
                         "Jingleheimer, Schmidt",
                         "JJJ, Schmidt"))
split_cols(data = d, col = "name", col_prefix = "names")
}
