% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_extract_matches.R
\name{str_extract_matches}
\alias{str_extract_matches}
\title{Match string and extract the matched value}
\usage{
str_extract_matches(string, pattern)
}
\arguments{
\item{string}{string to search}

\item{pattern}{regex pattern to match}
}
\value{
matched string
}
\description{
Match string and extract the matched value
}
\examples{
require(dplyr)
require(tidyr)
dplyr::mutate(dplyr::starwars,
      match = str_extract_matches(name, "Skywalker")) |>
  dplyr::select(last_col())
}
