% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ggp2g.R
\name{theme_ggp2g}
\alias{theme_ggp2g}
\title{ggplot2 theme (doping data)}
\usage{
theme_ggp2g(
  base_size = 11,
  base_family = "Ubuntu",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{size of font}

\item{base_family}{family of font}

\item{base_line_size}{size of lines in base graph}

\item{base_rect_size}{size of rectangle in base graph}
}
\value{
ggplot object
}
\description{
ggplot2 theme (doping data)
}
