% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sanctions_data.R
\name{util_read_cache}
\alias{util_read_cache}
\title{Read an object from the sanctions cache}
\usage{
util_read_cache(file)
}
\arguments{
\item{file}{cache file name, e.g. \code{"sanctions.rds"}}
}
\description{
Read an object from the sanctions cache
}
\keyword{internal}
