% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sanctions_data.R
\name{util_sanctions_cache_dir}
\alias{util_sanctions_cache_dir}
\title{Local cache directory for USADA sanctions data}
\usage{
util_sanctions_cache_dir()
}
\value{
path to the cache directory (created if missing)
}
\description{
Local cache directory for USADA sanctions data
}
\keyword{internal}
