% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sanctions_data.R
\name{util_write_cache}
\alias{util_write_cache}
\title{Write an object to the sanctions cache}
\usage{
util_write_cache(x, file)
}
\arguments{
\item{x}{object to cache}

\item{file}{cache file name, e.g. \code{"sanctions.rds"}}
}
\description{
Write an object to the sanctions cache
}
\keyword{internal}
