% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_inst_path.R
\name{verify_inst_path}
\alias{verify_inst_path}
\title{Verify inst/ path}
\usage{
verify_inst_path(inst_path = NULL)
}
\arguments{
\item{inst_path}{path string to validate as an \verb{inst/} folder path}
}
\value{
the validated path string, or an error if the path is not a valid
\verb{inst/} folder path.
}
\description{
Verify inst/ path
}
