% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_match_col.R
\name{add_match_col}
\alias{add_match_col}
\title{Create column with matched value (base R)}
\usage{
add_match_col(string, pattern)
}
\arguments{
\item{string}{string to search}

\item{pattern}{regex pattern to match}
}
\value{
matched string
}
\description{
Create column with matched value (base R)
}
\examples{
terms <- data.frame(term = c("A cramp is no small danger on a swim.",
                            "The soft cushion broke the man's fall.",
                             "There is a lag between thought and act.",
                             "Eight miles of woodland burned to waste."))
terms$match_upper <- add_match_col(terms$term, "[[:upper:]]")
terms$match_vowels <- add_match_col(terms$term, "[aeiou]")
terms
}
