% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polite-scrape.R
\name{check_rtxt}
\alias{check_rtxt}
\title{Function for checking robots.txt file}
\usage{
check_rtxt(url, delay, user_agent, force, verbose)
}
\arguments{
\item{url}{web address for download}

\item{delay}{default delay}

\item{user_agent}{user agent string}

\item{force}{force re-downloading of robots.xtx}

\item{verbose}{logical}
}
\description{
Function for checking robots.txt file
}
