% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_substance.R
\name{classify_substance}
\alias{classify_substance}
\title{Classify a specific WADA substance}
\usage{
classify_substance(df, subs_col, subs, var, val = "match", wb = TRUE)
}
\arguments{
\item{df}{data from USADA website}

\item{subs_col}{column with substances/reasons for sanctions}

\item{subs}{character string of banned WADA substances}

\item{var}{name of new variable}

\item{val}{variable value when matched}

\item{wb}{include word boundary}
}
\value{
substance dataset with newly classified substances
}
\description{
Classify a specific WADA substance
}
