% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_wada_substances.R
\name{classify_wada_substances}
\alias{classify_wada_substances}
\title{Classify banned WADA substances}
\usage{
classify_wada_substances(usada_data, subs_column)
}
\arguments{
\item{usada_data}{scraped data from USADA website}

\item{subs_column}{column with substances/reasons for sanctions}
}
\value{
substances dataset with newly classified substances
}
\description{
usada_data should be the table of sanctions from the
\link{United States Anti-Doping Agenc}y(https://www.usada.org/news/sanctions).
The \code{substance_reason} column contains the justification for each sanction.

\if{html}{\out{<div class="sourceCode">}}\preformatted{In some cases, there are multiple substances/reasons, and these should
be identified first.
}\if{html}{\out{</div>}}
}
\examples{
usada_data <- read.delim(
     system.file("extdata",
                 "raw",
                 "2023-12-19",
                 "2023-12-19-tidy_sports.csv", package = "dopingdata"),
               sep = ",")
substances <- classify_wada_substances(
  usada_data = usada_data,
  subs_column = substance_reason
)
head(substances[c('substance_group', 'substance_reason')])

}
