% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_wada_substances.R
\name{classify_wada_substances}
\alias{classify_wada_substances}
\title{Classify banned WADA substances}
\usage{
classify_wada_substances(usada_data, subs_column)
}
\arguments{
\item{usada_data}{scraped data from USADA website}

\item{subs_column}{column with substances/reasons for sanctions}
}
\value{
substances dataset with newly classified substances
}
\description{
usada_data should be the table of sanctions from the United
    States Anti-Doping Agency: https://www.usada.org/news/sanctions. The
    substance_reason column contains the justification for each sanction.

    In some cases, there are multiple substances/reasons, and these should
    be identified first.
}
\examples{
usada_data <- readr::read_csv(
  "inst/extdata/raw/2023-01-13/2023-01-13-usada_data_raw.csv"
)
classify_wada_substances(
  usada_data = usada_data,
  subs_column = substance_reason
)
}
