% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_dates.R
\name{clean_dates}
\alias{clean_dates}
\title{Clean sanction dates}
\usage{
clean_dates(df, date_col, split = "updated", pattern = "original")
}
\arguments{
\item{df}{processed USADA dataset with messy dates}

\item{date_col}{sanction date column (usually \code{sanction_announced})}

\item{split}{regex to pass to split argument of \code{strsplit()} (defaults to \code{"updated"})}

\item{pattern}{regex for other non-date pattern (defaults to \code{"original"})}
}
\value{
tibble with cleaned dates
}
\description{
Clean sanction dates
}
