% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_top_file.R
\name{clip_top_file}
\alias{clip_top_file}
\title{Copy the path to top data file in `extdata/` folder on the clipboard}
\usage{
clip_top_file(ext, pth, ctime)
}
\arguments{
\item{ext}{file extension. Default is `'csv'`, accepts: `'rds'`, `'txt'`, `'tsv'`,
or `'dat'`}

\item{pth}{path to `extdata` subfolder}
}
\value{
list with extension, path, and modification time for most recent file.
}
\description{
Copy the path to top data file in `extdata/` folder on the clipboard
}
