% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_extdata.R
\name{export_data}
\alias{export_data}
\title{Export raw data to path}
\usage{
export_data(x, path = "", type = "csv")
}
\arguments{
\item{x}{data to export}

\item{inst_path}{path to folder}
}
\value{
exported data
}
\description{
Export raw data to path
}
