% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recent.R
\name{get_recent}
\alias{get_recent}
\title{Return the most recent modification date}
\usage{
get_recent(pth = ".", full = FALSE)
}
\arguments{
\item{pth}{path to file or folder}

\item{full}{return datetime (instead of date)}
}
\value{
vrsn character vector of date or datetime
}
\description{
Return the most recent modification date
}
\examples{
get_recent("wrong")
get_recent("inst/extdata/")
get_recent("inst/extdata/raw", full = TRUE)
}
