% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recent_file.R
\name{get_recent_file}
\alias{get_recent_file}
\title{Return the most data file in folder}
\usage{
get_recent_file(path, type = "csv")
}
\arguments{
\item{path}{path to data folder}

\item{type}{file extension. Default is 'csv', accepts: 'rds', 'txt', 'tsv',
or 'dat'}
}
\value{
top_pth path the recent file
}
\description{
Return the most data file in folder
}
\examples{
require(fs)
path_files <- fs::dir_ls("inst/extdata/")
path_files
get_recent_file("inst/extdata/", "csv")
# create tmp dir
fs::dir_create("tmp/")
get_recent_file("tmp", "csv")
}
