% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recent_file.R
\name{get_recent_file}
\alias{get_recent_file}
\title{Return the most data file in folder}
\usage{
get_recent_file(path = "default", regex = NULL, ext = ".csv")
}
\arguments{
\item{path}{path to data folder}

\item{regex}{regular expression `pattern` passed to `list.files()`}

\item{ext}{file extension. Default is 'csv', accepts: 'rds', 'txt', 'tsv',
or 'dat'}
}
\value{
top_pth path the recent file
}
\description{
Return the most data file in folder
}
\examples{
get_recent_file(path = "extdata/raw", regex = "sanc", ext = "csv")
get_recent_file(regex = "subs", ext = "csv")
}
