% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polite-scrape.R
\name{guess_basename}
\alias{guess_basename}
\title{Guess filename for download from url}
\usage{
guess_basename(x)
}
\arguments{
\item{x}{url to guess filename from}
}
\description{
Guess filename for download from url
}
