% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_term_long.R
\name{pivot_term_long}
\alias{pivot_term_long}
\title{Parse string into individual terms (as tibble)}
\usage{
pivot_term_long(term, sep = "[^[:alnum:]]+")
}
\arguments{
\item{term}{}
}
\value{
tibble of unique terms and term
}
\description{
Parse string into individual terms (as tibble)
}
\examples{
pivot_term_long(term = "A large size in stockings is hard to sell.")
}
