% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polite-scrape.R
\name{polite_fetch_rtxt}
\alias{polite_fetch_rtxt}
\title{Function to get robots.txt is structured form. Memoised}
\usage{
polite_fetch_rtxt(..., user_agent, delay, verbose)
}
\arguments{
\item{...}{arguments passed to \code{robotstxt::robotstxt()}}

\item{user_agent}{user agent string}

\item{delay}{default delay}

\item{verbose}{logical}
}
\description{
Function to get robots.txt is structured form. Memoised
}
