% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_text.R
\name{process_text}
\alias{process_text}
\title{Process raw data}
\usage{
process_text(raw_data, fct = FALSE)
}
\arguments{
\item{raw_data}{raw dataset}

\item{fct}{include factors?}
}
\value{
tibble with standardized names, lowercase text, all character!
}
\description{
Process raw data
}
\examples{
require(palmerpenguins)
process_text(palmerpenguins::penguins_raw)
}
