% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_sanctions.R
\name{scrape_sanctions}
\alias{scrape_sanctions}
\title{Scrape the UASA website sanctions data}
\usage{
scrape_sanctions(dest_path)
}
\value{
tibble of raw sanctions data
}
\description{
Scrape the UASA website sanctions data
}
\examples{
# scrape_sanctions()
}
