% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_extdata.R
\name{verify_inst_path}
\alias{verify_inst_path}
\title{Verify inst/ path}
\usage{
verify_inst_path(inst_path = NULL)
}
\arguments{
\item{inst_path}{}
}
\value{
invisible
}
\description{
Verify inst/ path
}
